+++ src/theme/foundations/colors.ts
@@ -32,6 +32,8 @@ const grayMedium = "#BDBDBD";
 const grayLightCool = "#E9E9E9";
 const grayxLightCool = "#F5F5F5";
 const grayxxLightCool = "#FAFAFA";
+const linkPrimary = "#0576D3";
+const linkSecondary = "#004B98";
 const brandPrimary = "#C60917";
 const brandSecondary = "#760000";
 const blogsPrimary = grayLightCool;
@@ -142,8 +144,9 @@ const colors: Colors = {
     },
     focus: "#4181F1",
     link: {
-      primary: "#0576D3",
-      secondary: "#004B98",
+      primary: linkPrimary,
+      "primary-05": hexToRGB(linkPrimary, 0.05),
+      secondary: linkSecondary,
     },
     status: {
       primary: "#F9E08E",