+++ src/utils/utils.ts
@@ -1,6 +1,5 @@
 import { HelperErrorTextType } from "../components/HelperErrorText/HelperErrorText";
 import { AriaAttributes } from "./interfaces";
-
 // Utility functions to use throughout the codebase
 
 /**
@@ -94,4 +93,33 @@ export const hexToRGB = (hex: string, alpha: number) => {
     b = parseInt(bSlice, 16);
   const rgb = `${r}, ${g}, ${b}`;
   return alpha ? `rgba(${rgb},${alpha})` : `rgb(${rgb})`;
+};
+
+/** Calculate color luminance */
+export const colorLuminance = (r, g, b) => {
+  const a = [r, g, b].map(function (v) {
+    v /= 255;
+    return v <= 0.03928 ? v / 12.92 : Math.pow((v + 0.055) / 1.055, 2.4);
+  });
+  return a[0] * 0.2126 + a[1] * 0.7152 + a[2] * 0.0722;
+};
+
+/** Calculate color contrast ratio */
+export const contrastRatio = (hex1: string, hex2: string) => {
+  const rgb1 = hexToRGB(hex1, null);
+  const rgb2 = hexToRGB(hex2, null);
+  const colorLuminance1 = colorLuminance(rgb1[0], rgb1[1], rgb1[2]);
+  const colorLuminance2 = colorLuminance(rgb2[0], rgb2[1], rgb2[2]);
+  const ratio =
+    colorLuminance1 > colorLuminance2
+      ? (colorLuminance2 + 0.05) / (colorLuminance1 + 0.05)
+      : (colorLuminance1 + 0.05) / (colorLuminance2 + 0.05);
+
+  return (ratio * 10).toFixed(2);
+};
+
+/** This method will truncate text by centain length. Default length is 60. */
+export const truncateText = (text: string, truncateTextLength: number = 60) => {
+  const subString = text.substr(0, truncateTextLength - 1);
+  return `${subString.substr(0, subString.lastIndexOf(" "))}...`;
 };